change: AppServiceProvider::checkAuthority(), フレームがあれば、フレームを配置したページから親を遡ってページロールを取得に対応 add: ConnectRoleTrait, ConnectCommonTraitからRole関係を分離
