+++ app/Providers/AppServiceProvider.php
@@ -7,7 +7,7 @@ use Illuminate\Support\Facades\Gate;
 //use Illuminate\Support\ServiceProvider;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider;
 
-use App\Traits\ConnectCommonTrait;
+use App\Traits\ConnectRoleTrait;
 
 use App\Models\Common\Page;
 
@@ -16,7 +16,7 @@ use App\Enums\PluginName;
 //class AppServiceProvider extends ServiceProvider
 class AppServiceProvider extends AuthServiceProvider
 {
-    use ConnectCommonTrait;
+    use ConnectRoleTrait;
 
     /**
      * Register any application services.
@@ -402,8 +402,8 @@ class AppServiceProvider extends AuthServiceProvider
         $page_tree = $request->attributes->get('page_tree');
         // dd($page, $page->page_roles);
 
-        // 自分のページから親を遡ってページロールを取得
-        $page_roles = $this->getPageRolesByGoingBackParent($page, $page_tree);
+        // フレームがあれば、フレームを配置したページから親を遡ってページロールを取得
+        $page_roles = $this->choicePageRolesByGoingBackParentPageOrFramePage($page, $page_tree, $frame);
 
         // ユーザロール取得。所属グループのページ権限あったら、そっちからとる
         $user_roles = $this->choiceUserRolesOrPageRoles($user, $page_roles);
@@ -570,31 +570,8 @@ class AppServiceProvider extends AuthServiceProvider
         $page = $request->attributes->get('page');
         $page_tree = $request->attributes->get('page_tree');
 
-        // frame->page_id を基にページロール取得
-        // $page_roles = $page_roles ?? collect();
-        // $page_roles = $page_roles->where('page_id', $frame->page_id);
-        // \Log::debug(var_export($page_roles, true));
-
-        // \Log::debug('[' . __METHOD__ . '] ' . __FILE__ . ' (line ' . __LINE__ . ')');
-        // \Log::debug(var_export($frame->page_id, true));
-        // \Log::debug(var_export($page->id, true));
-
-        if ($page->id === $frame->page_id) {
-            // プラグインを配置したページと同じ（メインエリア等）
-            // 自ページのため、なにも変更しない
-
-        } else {
-            // プラグインを配置したページと違う（ヘッダーエリアや左エリア等）
-            // フレームの配置ページIDから、親を遡らせる。
-            $page = new Page();
-            $page->id = $frame->page_id;
-
-            // nullを指定する事で、フレームの配置ページから親を遡ってページツリーを再取得する。
-            $page_tree = null;
-        }
-
-        // ページから親を遡ってページロールを取得
-        $page_roles = $this->getPageRolesByGoingBackParent($page, $page_tree);
+        // フレームがあれば、フレームを配置したページから親を遡ってページロールを取得
+        $page_roles = $this->choicePageRolesByGoingBackParentPageOrFramePage($page, $page_tree, $frame);
 
 
         // 指定された権限を含むロールをループする。
